# apps/api/src/routes/rooms.ts
import type { FastifyInstance } from "fastify";
import { z } from "zod";
import { db } from "../db.js";
import { loadCurrentUser, requireAuth, requireRole } from "../middleware/auth.js";
import type { RoomListRow, RoomMessageRow, RoomRow } from "../db.types.ts";
import type { RoomCreateResponse, RoomMessagesResponse, RoomsListResponse } from "../api-contract.types.ts";

const createRoomSchema = z.object({
  slug: z
    .string()
    .min(3)
    .max(48)
    .regex(/^[a-z0-9-]+$/),
  title: z.string().min(3).max(120),
  is_public: z.boolean().default(true)
});

export async function roomsRoutes(fastify: FastifyInstance) {
  fastify.get(
    "/v1/rooms",
    {
      preHandler: [requireAuth]
    },
    async (request) => {
      const userId = String(request.user?.sub || "").trim();
      const result = await db.query<RoomListRow>(
        `SELECT
           r.id,
           r.slug,
           r.title,
           r.is_public,
           r.created_at,
           EXISTS(
             SELECT 1 FROM room_members rm
             WHERE rm.room_id = r.id AND rm.user_id = $1
           ) AS is_member
         FROM rooms r
         ORDER BY created_at ASC`,
        [userId]
      );

      const response: RoomsListResponse = { rooms: result.rows };
      return response;
    }
  );

  fastify.post<{ Body: { slug: string; title: string; is_public?: boolean } }>(
    "/v1/rooms",
    {
      preHandler: [requireAuth, loadCurrentUser, requireRole(["admin", "super_admin"])]
    },
    async (request, reply) => {
      const parsed = createRoomSchema.safeParse(request.body);

      if (!parsed.success) {
        return reply.code(400).send({
          error: "ValidationError",
          issues: parsed.error.flatten()
        });
      }

      const { slug, title, is_public } = parsed.data;

      const existing = await db.query("SELECT id FROM rooms WHERE slug = $1", [slug]);

      if ((existing.rowCount || 0) > 0) {
        return reply.code(409).send({
          error: "Conflict",
          message: "Room slug already exists"
        });
      }

      const createdBy = String(request.user?.sub || "").trim();

      const created = await db.query<RoomRow>(
        `INSERT INTO rooms (slug, title, is_public, created_by)
         VALUES ($1, $2, $3, $4)
         RETURNING id, slug, title, is_public, created_at`,
        [slug, title, is_public, createdBy]
      );

      const room = created.rows[0];

      await db.query(
        `INSERT INTO room_members (room_id, user_id, role)
         VALUES ($1, $2, 'owner')
         ON CONFLICT (room_id, user_id) DO NOTHING`,
        [room.id, createdBy]
      );

      const response: RoomCreateResponse = { room };
      return reply.code(201).send(response);
    }
  );

  fastify.get<{
    Params: { slug: string };
    Querystring: { limit?: string | number; beforeCreatedAt?: string; beforeId?: string };
  }>(
    "/v1/rooms/:slug/messages",
    {
      preHandler: [requireAuth]
    },
    async (request, reply) => {
      const userId = String(request.user?.sub || "").trim();
      const slug = String(request.params.slug || "").trim();
      const limit = Math.min(100, Math.max(1, Number(request.query.limit || 50)));
      const beforeCreatedAtRaw = String(request.query.beforeCreatedAt || "").trim();
      const beforeIdRaw = String(request.query.beforeId || "").trim();

      let beforeCreatedAt: string | null = null;
      let beforeId: string | null = null;

      if (beforeCreatedAtRaw || beforeIdRaw) {
        if (!beforeCreatedAtRaw || !beforeIdRaw) {
          return reply.code(400).send({
            error: "ValidationError",
            message: "beforeCreatedAt and beforeId must be provided together"
          });
        }

        const beforeDate = new Date(beforeCreatedAtRaw);
        if (Number.isNaN(beforeDate.getTime())) {
          return reply.code(400).send({
            error: "ValidationError",
            message: "beforeCreatedAt must be a valid ISO datetime"
          });
        }

        beforeCreatedAt = beforeDate.toISOString();
        beforeId = beforeIdRaw;
      }

      const roomResult = await db.query<RoomRow>(
        "SELECT id, slug, title, is_public FROM rooms WHERE slug = $1",
        [slug]
      );

      if (roomResult.rowCount === 0) {
        return reply.code(404).send({
          error: "RoomNotFound",
          message: "Room does not exist"
        });
      }

      const room = roomResult.rows[0];

      if (!room.is_public) {
        const membership = await db.query(
          "SELECT 1 FROM room_members WHERE room_id = $1 AND user_id = $2",
          [room.id, userId]
        );

        if (membership.rowCount === 0) {
          return reply.code(403).send({
            error: "Forbidden",
            message: "You cannot access this room"
          });
        }
      }

      const messagesResult = beforeCreatedAt && beforeId
        ? await db.query<RoomMessageRow>(
            `SELECT
               m.id,
               m.room_id,
               m.user_id,
               m.body AS text,
               m.created_at,
               u.name AS user_name
             FROM messages m
             JOIN users u ON u.id = m.user_id
             WHERE m.room_id = $1
               AND (m.created_at, m.id) < ($2::timestamptz, $3)
             ORDER BY m.created_at DESC, m.id DESC
             LIMIT $4`,
            [room.id, beforeCreatedAt, beforeId, limit + 1]
          )
        : await db.query<RoomMessageRow>(
            `SELECT
               m.id,
               m.room_id,
               m.user_id,
               m.body AS text,
               m.created_at,
               u.name AS user_name
             FROM messages m
             JOIN users u ON u.id = m.user_id
             WHERE m.room_id = $1
             ORDER BY m.created_at DESC, m.id DESC
             LIMIT $2`,
            [room.id, limit + 1]
          );

      const hasMore = messagesResult.rows.length > limit;
      const pageDesc = hasMore ? messagesResult.rows.slice(0, limit) : messagesResult.rows;
      const oldestInPage = pageDesc[pageDesc.length - 1] || null;

      const response: RoomMessagesResponse = {
        room,
        messages: pageDesc.reverse(),
        pagination: {
          hasMore,
          nextCursor: hasMore && oldestInPage
            ? {
                beforeCreatedAt: oldestInPage.created_at,
                beforeId: oldestInPage.id
              }
            : null
        }
      };
      return response;
    }
  );
}