# docs/OPENAPI_V1.yaml
openapi: 3.0.3
info:
  title: Boltorezka API
  version: 1.0.0
  description: >-
    Canonical HTTP contract for Boltorezka v1.
    This spec tracks currently implemented endpoints.
servers:
  - url: https://test.boltorezka.gismalink.art
    description: Test
  - url: https://boltorezka.gismalink.art
    description: Production
  - url: http://localhost:8080
    description: Local
...
paths:
...
  /v1/rooms:
    get:
      tags: [rooms]
      summary: List rooms visible to current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Rooms list
          content:
            application/json:
              schema:
                type: object
                properties:
                  rooms:
                    type: array
                    items:
                      $ref: '#/components/schemas/RoomListItem'
    post:
      tags: [rooms]
      summary: Create room (admin/super_admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slug, title]
              properties:
                slug:
                  type: string
                title:
                  type: string
                is_public:
                  type: boolean
      responses:
        '201':
          description: Room created
          content:
            application/json:
              schema:
                type: object
                properties:
                  room:
                    $ref: '#/components/schemas/Room'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/rooms/{slug}/messages:
    get:
      tags: [rooms]
      summary: Get room message history
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - in: query
          name: beforeCreatedAt
          schema:
            type: string
            format: date-time
        - in: query
          name: beforeId
          schema:
            type: string
      responses:
        '200':
          description: Messages in room
          content:
            application/json:
              schema:
                type: object
                properties:
                  room:
                    $ref: '#/components/schemas/Room'
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/RoomMessage'
                  pagination:
                    type: object
                    properties:
                      hasMore:
                        type: boolean
                      nextCursor:
                        type: object
                        nullable: true
                        properties:
                          beforeCreatedAt:
                            type: string
                            format: date-time
                          beforeId:
                            type: string
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/RoomNotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Room:
      type: object
      properties:
        id: { type: string }
        slug: { type: string }
        title: { type: string }
        is_public: { type: boolean }
        created_at:
          type: string
          format: date-time
    RoomListItem:
      allOf:
        - $ref: '#/components/schemas/Room'
        - type: object
          properties:
            is_member:
              type: boolean
    RoomMessage:
      type: object
      properties:
        id: { type: string }
        room_id: { type: string }
        user_id: { type: string }
        text: { type: string }
        user_name: { type: string }
        created_at:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
        message: { type: string }
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    RoomNotFound:
      description: Room not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...